requirements.txt
altgraph==0.17.2
imap-tools==0.49.1
macholib==1.15.2
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
python-dotenv==0.19.1
